:hammer: Ajoute les contrats Madelin pour l'option IS
